feat: mocking data in schemas
This change for division in schemas between ID and UUID scalars which we use in type and interface. In addition we use fake directives in type scalars, change scalar type ID to UUID, String to DateTime in insertedAt and updatedAt field on interface contracts and contractRequests, delete fake directives in Date and DateTime scalars.

BREAKING CHANGE: change scalar type ID to UUID in schemas